+++ lib/quickProps/index.ts
@@ -1,2 +1,0 @@
-export { dimensions } from './dimensions'
-export { spacing } from './spacing'