+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 ## Changelog:
 
-0.0.412:
-Breaking Changes
+### 0.0.6
+- Updated to use styled-systems
+
+*Breaking Changes*
+- s.media.w(123) becomes ${s.mediaQuery(120)} { /* CSS */ }
+
+### 0.0.412:
+*Breaking Changes*
 - changed mediaDimension to breakpoints
 - Changed how media queries are used
 - added array types to margin and padding variables